move v0.1.0 code to reference directory
