+++ _v1/PALAS_USMPS_fileread.py
@@ -20,9 +20,9 @@ from datetime import datetime
 
 import numpy as np
 import pandas as pd
-
-from defs import device_list
 from sup import convert_mbar_to_kPa, get_filename
+
+from _v1.defs import device_list
 
 
 def import_data(filename, data_choice=""):
