+++ _v1/TSI_SMPS_fileread.py
@@ -14,9 +14,9 @@ from datetime import datetime
 
 import numpy as np
 import pandas as pd
-
-from defs import device_list
 from sup import convert_kPa_to_mbar, get_filename
+
+from _v1.defs import device_list
 
 
 def rename_columns(df, used_device):
